refactor: move the defaults setup to the cart loading
