+++ client-app/pages/checkout/checkout.vue
@@ -189,7 +189,6 @@ import { usePopup } from "@/shared/popup";
 import { computed, onBeforeUpdate, onMounted, ref } from "vue";
 import _ from "lodash";
 import { PaymentMethodType, ShippingMethodType } from "@/core/api/graphql/types";
-import { useUserCheckoutDefaults } from "@/shared/account";
 
 const {
   cart,
@@ -206,8 +205,6 @@ const {
   updatePayment,
   loading,
 } = useCart();
-
-const { getUserCheckoutDefaults } = useUserCheckoutDefaults();
 
 const { placeOrder } = useCheckout();
 const { openPopup } = usePopup();
@@ -295,26 +292,6 @@ onMounted(async () => {
       cartCouponApplied.value = true;
     }
     cartComment.value = cart.value.comment || "";
-
-    //#region set checkout defaults
-    const checkoutDefaults = getUserCheckoutDefaults();
-    if (!cart.value.shipments?.[0]?.id && checkoutDefaults?.shippingMethod) {
-      const method = checkoutDefaults?.shippingMethod;
-      await updateShipment({
-        shipmentMethodCode: method.code,
-        shipmentMethodOption: method.optionName,
-        id: cart.value.shipments?.[0]?.id,
-      });
-    }
-
-    if (!cart.value.payments?.[0]?.id && checkoutDefaults?.paymentMethod) {
-      const method = checkoutDefaults?.paymentMethod;
-      await updatePayment({
-        paymentGatewayCode: method.code,
-        id: cart.value.payments?.[0]?.id,
-      });
-    }
-    //#endregion set checkout defaults
   });
 });
